use f64 instead of f32 for compute final score

+++ src/gradualsemantics/categorizer.rs
@@ -1,14 +1,14 @@
 use crate::{graph::ArgumentationFramework, parser::Task};
 
-const EPSILON : f32 = 0.0001;
+const EPSILON : f64 = 0.0001;
 
-pub fn solve(af : ArgumentationFramework, task : &Task) -> f32 {
+pub fn solve(af : ArgumentationFramework, task : &Task) -> f64 {
     let score = computeFinalScore(&af);
     let solution = score[task.argument];
     solution
 }
 
-fn computeFinalScore(af : &ArgumentationFramework) -> Vec<f32> {
+fn computeFinalScore(af : &ArgumentationFramework) -> Vec<f64> {
     let mut res = initScores(af);
     let mut hasChanged = true;
 		while hasChanged {
@@ -21,7 +21,7 @@ fn computeFinalScore(af : &ArgumentationFramework) -> Vec<f32> {
 		return res;
 }
 
-fn computeOneStep(af : &ArgumentationFramework, scoresArg : &Vec<f32>) -> Vec<f32> {
+fn computeOneStep(af : &ArgumentationFramework, scoresArg : &Vec<f64>) -> Vec<f64> {
     let mut res = vec![0.;scoresArg.len()];
 		for i in 0..scoresArg.len() {
 			let mut sumScoreAttacker = 0.;
@@ -33,11 +33,11 @@ fn computeOneStep(af : &ArgumentationFramework, scoresArg : &Vec<f32>) -> Vec<f3
 		return res;
 }
 
-fn initScores(af : &ArgumentationFramework) -> Vec<f32> {
+fn initScores(af : &ArgumentationFramework) -> Vec<f64> {
     vec![1.0;af.nb_argument]
 }
 
-fn stabilisation(tab1 : &Vec<f32>, tab2 : &Vec<f32>) -> bool {
+fn stabilisation(tab1 : &Vec<f64>, tab2 : &Vec<f64>) -> bool {
 	for (i, x) in tab1.into_iter().enumerate() {
 		if (x-tab2[i]).abs() > EPSILON {
 			return false;
